+++ ChangeLog.md
@@ -1,5 +1,21 @@
 # Revision history for jak
 
-## 0.1.0.0  -- YYYY-mm-dd
+## 0.3.0.0 -- 2018-02-01
+
+* Properly implement event handling with frp,
+  now every component handles its own events.
+* Split DoubleSeq.hs.
+* Enable mouse events.
+* Moved design decisions to README.
+* Write this ChangeLog.
+
+## 0.2.0.0 -- 2018-01-18
+
+* Split DoubleSeq into Editor, Viewport, Cursor and Content.
+* Remove outdated SingleSeq.
+* Jak now has a viewport.
+
+## 0.1.0.0  -- 2018-01-07
 
 * First version. Released on an unsuspecting world.
+* Jak can edit text now!